refactor type defs for account/invoice into their own files. refactored how event types work so there are now subtypes. Added first pass 'settings' for payout logic, however need to chat with Raffe about refactoring account upsert logic in API as it's inconsistent throughout the webapi -> api -> store layers.
