+++ pkg/events.go
@@ -67,6 +67,7 @@ const (
 	INV_CREATED          EVENT_INV = "CREATED"
 	INV_UPDATED          EVENT_INV = "UPDATED"
 	INV_PAYMENT_RECEIVED EVENT_INV = "PAYMENT_RECEIVED"
+	INV_PAYMENT_SENT     EVENT_INV = "PAYMENT_SENT"
 	INV_PAYMENT_VERIFIED EVENT_INV = "PAYMENT_VERIFIED"
 	INV_PAYMENT_REFUNDED EVENT_INV = "PAYMENT_REFUNDED"
 )
