+++ pkg/store.go
@@ -3,6 +3,7 @@ package giga
 type Store interface {
 	// StoreInvoice stores an invoice.
 	StoreInvoice(invoice Invoice) error
+
 	// GetInvoice returns the invoice with the given ID.
 	GetInvoice(id Address) (Invoice, error)
 	// ListInvoices returns a filtered list of invoices for an account.
@@ -10,11 +11,13 @@ type Store interface {
 	// pagination: when next_cursor == 0, that is the final page of results.
 	// pagination: stores CAN return < limit (or zero) items WITH next_cursor > 0 (due to filtering)
 	ListInvoices(account Address, cursor int, limit int) (items []Invoice, next_cursor int, err error)
+
 	// GetPendingInvoices sends all invoices that are pending to the given channel.
 	GetPendingInvoices() (<-chan Invoice, error)
 
 	// StoreAccount stores an account.
 	StoreAccount(account Account) error
+
 	// GetAccount returns the account with the given ForeignID.
 	GetAccount(foreignID string) (Account, error)
